Refine file structures for deployment
