+++ Procfile
@@ -1,1 +1,1 @@
-web: node server/index.js
+web: node server.js
